bruce-web-test/requirement.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
paramiko==3.4.0
aiohttp==3.9.1
PyYAML==6.0.1
pydantic==2.5.0
numpy==1.26.0
pandas==2.1.4
matplotlib==3.8.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-multipart==0.0.6